requirements.txt
python-dotenv==1.0.1
ccxt==4.3.58
pandas==2.2.3
numpy==1.26.4
streamlit==1.41.1
plotly==5.24.1